P4ADEV-320-refactoring-accordingly-handbook
