+++ src/test/java/it/gov/pagopa/payhub/auth/exception/AuthExceptionHandlerTest.java
@@ -17,6 +17,7 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import static org.mockito.Mockito.doThrow;
@@ -28,6 +29,7 @@ import static org.mockito.Mockito.doThrow;
         AuthExceptionHandler.class})
 class AuthExceptionHandlerTest {
 
+    public static final String DATA = "data";
     @Autowired
     private MockMvc mockMvc;
 
@@ -39,16 +41,17 @@ class AuthExceptionHandlerTest {
     static class TestController {
 
         @GetMapping("/test")
-        String testEndpoint() {
+        String testEndpoint(@RequestParam(DATA) String data) {
             return "OK";
         }
     }
 
     @Test
     void handleInvalidTokenException() throws Exception {
-        doThrow(new InvalidTokenException("Error")).when(testControllerSpy).testEndpoint();
+        doThrow(new InvalidTokenException("Error")).when(testControllerSpy).testEndpoint(DATA);
 
         mockMvc.perform(MockMvcRequestBuilders.get("/test")
+                        .param(DATA, DATA)
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
                 .andExpect(MockMvcResultMatchers.status().isUnauthorized())
@@ -59,14 +62,27 @@ class AuthExceptionHandlerTest {
 
     @Test
     void handleTokenExpiredException() throws Exception {
-        doThrow(new TokenExpiredException("Error")).when(testControllerSpy).testEndpoint();
+        doThrow(new TokenExpiredException("Error")).when(testControllerSpy).testEndpoint(DATA);
 
         mockMvc.perform(MockMvcRequestBuilders.get("/test")
+                        .param(DATA, DATA)
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
                 .andExpect(MockMvcResultMatchers.status().isUnauthorized())
                 .andExpect(MockMvcResultMatchers.jsonPath("$.code").value("AUTH_TOKEN_EXPIRED_DATE"))
                 .andExpect(MockMvcResultMatchers.jsonPath("$.message").value("Error"));
+
+    }
+
+    @Test
+    void handleMissingServletRequestParameterException() throws Exception {
+
+        mockMvc.perform(MockMvcRequestBuilders.get("/test")
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .accept(MediaType.APPLICATION_JSON))
+                .andExpect(MockMvcResultMatchers.status().isBadRequest())
+                .andExpect(MockMvcResultMatchers.jsonPath("$.code").value("AUTH_INVALID_REQUEST"))
+                .andExpect(MockMvcResultMatchers.jsonPath("$.message").value("Required request parameter 'data' for method parameter type String is not present"));
 
     }
 }
